feat(model): Added Cryptol theories for initial version of cryptographic primitives, fulfilling the completion criteria in Issues 156 & 157 as well as PR 231. This includes:
(1) Pseudo-Random Number Generation from Decision Diffie-Hellman (DDH) Assumption, Section 10.1 of Doug's EVS Draft.

(2) ElGamal Cryptosystem (Section 10.4 of the EVS Draft) including functions for homomorphic mixing.

(3) Monadic lifting of ElGamal to illustrate use of state/random monads, in case we later want to adopt this approach in the Rust API for our library.

(4) The k-Pedersen Commitment Scheme (Section 10.3 of the EVS Draft), also including the simple version of Pedersen. Seeded variants of functions are provided that construct all randomness based on a single seed.

(5) Instantiation of (2) to (4) with several concrete groups, including a simple additive group module N, Schnorr Groups, and Prime Factor Elliptic Curve (PFEC) groups provided by cryptol-specs.

(6) Verification of all of the above either via formal proof or (model-)checking in Cryptol, as it is feasible.

(7) Rust code generation, compilation and execution of a series of smoke tests for (1) to (4), considering all instances in (5). We debugged and worked around any emerging Cryptol to Rust compilation and runtime issues, due to the structure of specifications.

(8) Making sure that verification and smoke-testing activities succeed for _all_ Cryptol modules under Primitives and are moreover executed by the CI pipeline.