chore: eliminate Cryptol symlinks and update CI
This commit removes the symlink from `Common/utils.cry` into the
`cryptol-specs` submodule, replacing it with a copy of the actual
Cryptol file, to make building more straightforward (especially on
Windows, where Git does not support symlinks in any reasonable way).
It updates the Rust codegen CI to account for this change, and also
to behave reasonably when run on a version of the repo that doesn't
support Rust code generation.

It also updates cryptol-specs to the latest version, though we aren't
currently using any of it except for `Common/utils.cry`. If that changes,
we will either copy the files we use into our tree, or change our reuse
strategy (e.g., to use CRYPTOLPATH to load the `cryptol-specs` library).